pyproject.toml
[project]
name = "ai-test-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]
